cleanup(ionic-react): refactor Jest generation
This change refactors the way that Jest files are made entirely. First of all, the `@nrwl/react:application` schematic is executed with no test runner so that we can have complete control over how the Jest schematic is used. Second of all, we look at the generated apps workspace to find the Jest config file path and update the file instead of overwrite it. This will be much more future proof in case Nrwl makes changes to the Jest configs contents or location.