chore(ionic-react): refactor templates and tests
This change came out of working on adding the other Ionic starter templates. The complexity of supporting multiple templates seems to be too much for what its worth, but there were some things that came out of that work that can be salvaged. The templates have been refactored to be much more readable, and the tests now have much wider coverage.