fix(ionic-react): properly initialize Capacitor plugin
The `@nxtend/capacitor` plugin is not properly initialized with the `@nxtend/ionic-react:init` schematic. This results in an error being thrown if an Ionic React application is attempted to be generated if `@nxtend/capacitor` has not been added to the users workspace manually.

This change ensures that the `@nxtend/capacitor` plugin is added to the users `package.json` and is initialized before generating an application with Capacitor enabled.

Unit and e2e tests have also been added to ensure this functionality doesn't break in the future.