fix(ionic-react): add @nrwl/react version based on the users Nx version
When the `@nxtend/ionic-react:init` schematic is executed and the user does not have the `@nrwl/react` plugin installed then it will add the `@nrwl/react` dependency with the latest version.

This change ensures that the version of `@nrwl/react` being added matches their `@nrwl/workspace` version to avoid unwanted side effects.