feat(capacitor): copy package.json from workspace root for cap commands (#204)
Capacitor requires a `package.json` in the folder where the binary is being executed. In this case, it is the sub-directory where the Capacitor project was generated. A second `package.json` was added while generating the Capacitor project to get around this issue, however, it required all Capacitor plugins to be added to both the root and the project-level `package.json`. This change removes the need for the project-level `package.json` by copying the file from the root of the workspace temporarily.